Moved camera registry to registries
